+++ packages/components/src/Icon/Icon.vue
@@ -4,12 +4,8 @@
  */
 import { toRef } from "vue";
 
-import type { IconName } from "@knime/kds-styles/img/icons/def";
-
-import { type IconSize } from "./types";
+import type { IconName, IconSize } from "./types";
 import useIcon from "./useIcon";
-
-export type { IconName, IconSize };
 
 const props = withDefaults(defineProps<{ name: IconName; size?: IconSize }>(), {
   size: "medium",